+++ app/src/main/res/layout/fragment_sharebottomsheet.xml
@@ -23,21 +23,21 @@
             android:layout_marginTop="100dp">
 
             <ImageButton
-                android:id="@+id/shareBottomSheet_image_btn"
+                android:id="@+id/shareBottomSheet_logo_btn"
                 android:layout_width="70dp"
                 android:layout_height="70dp"
                 android:layout_marginLeft="50dp"
                 android:background="@drawable/logo"/>
 
             <ImageButton
-                android:id="@+id/shareBottomSheet_char_btn"
+                android:id="@+id/shareBottomSheet_image_btn"
                 android:layout_width="70dp"
                 android:layout_height="70dp"
                 android:layout_marginLeft="50dp"
                 android:background="@drawable/myprofile_image"/>
 
             <ImageButton
-                android:id="@+id/shareBottomSheet_logo_btn"
+                android:id="@+id/shareBottomSheet_char_btn"
                 android:layout_width="70dp"
                 android:layout_height="70dp"
                 android:layout_marginLeft="50dp"
